Fix docker command not implemented in NodeJS image Fix path for docker usage Fix docker-compose build Update Node version to v18.4.0

+++ start.sh
@@ -1,4 +1,4 @@
-#! /bin/ash
+#! /bin/sh
 
 if [ ! -d "config" ]; then
     echo "Creating config dir"
@@ -6,6 +6,10 @@ if [ ! -d "config" ]; then
 fi
 
 if [ ! -f "config/default.json" ]; then
+    if [ ! -f "sample_config/default.json" ]; then
+        echo "No default config found, exiting"
+        exit 1
+    fi
     echo "Copying default config"
     cp sample_config/default.json config
 fi
